WEBSCRAPE-0011 move backend scripts to utils dir and create links in the root dir
